+++ src/FluxStudisSelfserviceFrontend.mjs
@@ -1,6 +1,5 @@
 import { flux_css_api } from "./Libs/flux-css-api/src/FluxCssApi.mjs";
 import { HttpClientResponse } from "./Libs/flux-http-api/src/Client/HttpClientResponse.mjs";
-import { COLOR_SCHEME_DARK, COLOR_SCHEME_LIGHT } from "./Libs/flux-color-scheme/src/ColorScheme/COLOR_SCHEME.mjs";
 import { PAGE_CHOICE_SUBJECT, PAGE_COMPLETED, PAGE_CREATE, PAGE_IDENTIFICATION_NUMBER, PAGE_INTENDED_DEGREE_PROGRAM, PAGE_INTENDED_DEGREE_PROGRAM_2, PAGE_LEGAL, PAGE_PERSONAL_DATA, PAGE_PORTRAIT, PAGE_PREVIOUS_STUDIES, PAGE_RESUME, PAGE_START, PAGE_UNIVERSITY_ENTRANCE_QUALIFICATION } from "./Page/PAGE.mjs";
 import { SETTINGS_INDEXEDDB_IMPLEMENTATION_DATABASE_NAME, SETTINGS_INDEXEDDB_IMPLEMENTATION_STORE_NAME } from "./Settings/SETTINGS_IMPLEMENTATION.mjs";
 
@@ -203,23 +202,9 @@ export class FluxStudisSelfserviceFrontend {
      * @returns {Promise<FluxColorScheme>}
      */
     async #getFluxColorScheme() {
-        this.#flux_color_scheme ??= (await import("./Libs/flux-color-scheme/src/FluxColorScheme.mjs")).FluxColorScheme.new(
-            [
-                {
-                    color_scheme: COLOR_SCHEME_LIGHT,
-                    name: "light"
-                },
-                {
-                    color_scheme: COLOR_SCHEME_DARK,
-                    name: "dark"
-                }
-            ],
-            await this.#getFluxLocalizationApi(),
-            await this.#getFluxSettingsApi(),
-            {
-                [COLOR_SCHEME_LIGHT]: "light",
-                [COLOR_SCHEME_DARK]: "dark"
-            },
+        this.#flux_color_scheme ??= await (await import("./Libs/flux-color-scheme/src/FluxColorScheme.mjs")).FluxColorScheme.new(
+            null,
+            null,
             [
                 "container-border-color",
                 "form-background-color",
@@ -227,7 +212,12 @@ export class FluxStudisSelfserviceFrontend {
                 "input-border-color",
                 "left-background-color",
                 "left-border-color"
-            ]
+            ],
+            null,
+            null,
+            null,
+            await this.#getFluxLocalizationApi(),
+            await this.#getFluxSettingsApi()
         );
 
         return this.#flux_color_scheme;